Massive classes renaming Adding dev docs
