+++ searchf/views.py
@@ -1,111 +1,28 @@
 '''Module implementing views for searchf application.'''
 
-import copy
 import curses
 import curses.ascii
 import os
 import re
 import sys
 
-from curses.textpad import Textbox
-from typing import Any
 from typing import List
 from typing import NamedTuple
 from typing import Optional
 from typing import Tuple
 
-from . import __version__
 from . import colors
-from . import debug
 from . import enums
 from . import models
 from . import segments
 from . import types
 
-# Changes layout to show a debug window in which debug() function will output
-USE_DEBUG = False
 
 # Do not use curses.A_BOLD on Windows as it just renders horribly when
 # highlighting
 USE_BOLD = 0 if sys.platform == 'win32' else curses.A_BOLD
 
 STATUS_EMPTY = ''
-STATUS_UNCHANGED = 'unchanged'
-
-
-def _load_help_lines():
-    lines = [f'  ~ Searchf {__version__} Help ~', '']
-    help_file = os.path.join(
-        os.path.dirname(os.path.abspath(__file__)),
-        'help.txt')
-    with open(help_file, encoding='utf-8') as f:
-        lines += f.readlines()
-    return lines
-
-
-HELP_LINES = _load_help_lines()
-
-
-def get_max_yx(scr) -> types.Size:
-    '''This function is a test artifact that wraps getmaxyx() from curses so
-    that we can overwrite it and test specific dimensions.
-    '''
-    return scr.getmaxyx()
-
-
-def getmtime(path):
-    '''Wraps os.getmtime() for testing'''
-    return os.path.getmtime(path)
-
-
-class EscapeException(Exception):
-    '''Signals that Escape key has been pressed'''
-
-
-def validate(k: int) -> int:
-    '''Validates key'''
-    if k == curses.ascii.DEL:
-        k = curses.KEY_BACKSPACE
-    elif k == curses.ascii.ESC:
-        raise EscapeException()
-    return k
-
-
-def get_text(scr, y, x, text_prompt: str, handler, text: str) -> str:
-    '''Gets text interactively from end user'''
-    scr.addstr(y, x, text_prompt)
-    x += len(text_prompt)
-    editwin = curses.newwin(1, 30, y, x)
-    scr.refresh()
-    box = Textbox(editwin)
-    for c in text:
-        box.do_command(ord(c))
-    text = ''
-    try:
-        handler(box)
-        text = box.gather().strip()
-    except EscapeException:
-        pass
-
-    editwin.clear()
-    editwin.refresh()
-    clear(scr, y, 0, len(text_prompt))
-    return text if text else ''
-
-
-def prompt(scr, y: int, x: int, text_prompt: str, text: str) -> str:
-    '''Prompts user to enter some text.'''
-    def handle(box):
-        box.edit(validate=validate)
-    return get_text(scr, y, x, text_prompt, handle, text)
-
-
-def clear(scr, y, x, length):
-    '''Prints "length" spaces at the given position'''
-    _, maxw = get_max_yx(scr)
-    blank = f'{" ":>{length}}'
-    scr.addstr(y, x, blank[:maxw-(x+1)])
-    scr.move(y, x)
 
 
 def bool_to_text(value: bool) -> str:
@@ -141,16 +58,21 @@ class TextView:
     # pylint: disable=too-many-instance-attributes
 
     def __init__(self, store, scr, name: str, path: str) -> None:
-        self._model: models.Model = models.Model()
-        self._vm: models.ViewModel = models.ViewModel()
+        self._raw: models.RawContent = \
+            models.RawContent()
+        self._selected: models.SelectedContent = \
+            models.SelectedContent()
+        self._display: models.DisplayContent = \
+            models.DisplayContent()
+        self._offsets: models.Offsets = models.Offsets()
         self._config: models.ViewConfig = models.ViewConfig()
         self._store = store
         self._scr = scr
         self._name: str = name
         self._basename = os.path.basename(path)
-        self._max_visible_lines_count = 0
         self._win: Optional[curses._CursesWindow] = None
         self._size: types.Size = (0, 0)
+        self._content_available_size: types.Size = (0, 0)
         self._ruler: str = ''
 
     def get_config(self) -> models.ViewConfig:
@@ -202,7 +124,7 @@ class TextView:
 
     def _get_prefix_info(self) -> PrefixInfo:
         if self._config.line_numbers:
-            number_length = self._model.line_number_length()
+            number_length = self._raw.line_number_length()
             sep = ' │ '
         elif self._config.wrap:
             number_length = 0
@@ -225,7 +147,8 @@ class TextView:
         if not self._config.has_filters():
             self._win.addstr(' No filter ', style)
         else:
-            self._win.addstr(f'{self._model.hits_count():>8}', style)
+            self._win.addstr(f'{self._selected.hits_count():>8}',
+                             style)
             self._win.addstr(f' | {"Case":^{CASE_COL_LEN}}', style)
             self._win.addstr(f' | {"Shown":^{SHOWN_COL_LEN}}', style)
             self._win.addstr(' | Keywords ', style)
@@ -242,11 +165,11 @@ class TextView:
         # voffest_desc is not always shown, but at most 5 char long
         text = ''.ljust(5)
         x = move_left_for(x, text)
-        if len(self._vm.voffset_desc) > 0:
-            text = f' {self._vm.voffset_desc:>3} '
+        if len(self._offsets.voffset_desc) > 0:
+            text = f' {self._offsets.voffset_desc:>3} '
             self._win.addstr(y, x, text, style)
 
-        text = f' {self._model.line_count()} lines '
+        text = f' {self._raw.line_count()} lines '
         x = move_left_for(x, text)
         self._win.addstr(y, x, text, style)
 
@@ -254,11 +177,12 @@ class TextView:
         x = move_left_for(x, text)
         self._win.addstr(y, x, text, style | USE_BOLD)
 
-        assert len(self._model.hits) == self._config.get_filters_count()
+        assert len(self._selected.hits) == \
+            self._config.get_filters_count()
         for i, f in enumerate(self._config.filters):
             x = 0
             y += 1
-            self._win.addstr(y, 0, f'{self._model.hits[i]:>8}')
+            self._win.addstr(y, 0, f'{self._selected.hits[i]:>8}')
             text = CASE_COL_TEXTS[1 if f.ignore_case else 2]
             self._win.addstr(f' | {text:^{CASE_COL_LEN}}')
             text = SHOWN_COL_TEXTS[1 if f.hiding else 2]
@@ -276,42 +200,47 @@ class TextView:
 
     def _draw_content(self, position, text, matching_segments, offset, color):
         y, x = position
-        vend = offset + self._vm.size[1]
+        _, width = self._content_available_size
+        vend = offset + width
         if self._config.show_spaces:
             text = text.replace(' ', '·')  # Note: this is curses.ACS_BULLET
         if self._config.colorize_mode == enums.ColorizeMode.LINE:
             text = text[offset:vend]
-            text = f'{text:<{self._vm.size[1]}}'
+            text = f'{text:<{width}}'
             self._win.addnstr(y, x, text, self._size[1], color)
         else:
             for match, start, end in segments.iterate(
                     offset, vend, matching_segments):
                 assert start < end
-                c = color if match else 0
                 length = end - start
-                self._win.addnstr(y, x, text[start:end], length, c)
+                self._win.addnstr(y, x,
+                                  text[start:end],
+                                  length,
+                                  color if match else 0)
                 x += length
 
     def draw(self):
         '''Draws the view'''
-        # debug.out(f'{self._name} draw {self._vm.voffset}')
+        # debug.out(f'{self._name} draw {self._offsets.voffset}')
         self._win.clear()
 
         prefix_info = self._get_prefix_info()
+        ymax, _ = self._content_available_size
 
         # Only draw what we need: we are going to pull at most
-        # _max_visible_lines_count lines from the data model, possibly less if
-        # some content needs multiple lines (wrapping)
-        iddata = self._vm.voffset
-        iddatamax = self._vm.lines_count()
+        # _content_available_size[1] lines from DisplayContent, possibly
+        # less if some content needs multiple lines (wrapping)
 
-        for y in range(self._max_visible_lines_count):
-            if iddata >= iddatamax:
+        idline = self._offsets.voffset
+        idlinemax = self._display.lines_count()
+
+        for y in range(ymax):
+            if idline >= idlinemax:
                 break
-            idata, offset = self._vm.data[iddata]
-            iddata += 1
+            iline, offset = self._display.dlines[idline]
+            idline += 1
             line_idx, filter_idx, text, matching_segments = \
-                self._model.data[idata]
+                self._selected.lines[iline]
             if line_idx == models.RULER_INDEX:
                 self._win.addstr(y, 0, self._ruler)
                 continue
@@ -321,7 +250,7 @@ class TextView:
             # the prefix on the first line, which has offset 0.
             if offset == 0:
                 self._draw_prefix(y, prefix_info, line_idx, color)
-            offset += self._vm.hoffset
+            offset += self._offsets.hoffset
             self._draw_content(
                 (y, prefix_info.length),
                 text,
@@ -329,37 +258,41 @@ class TextView:
                 offset,
                 color)
 
-        self._draw_bar(self._max_visible_lines_count)
+        self._draw_bar(self._content_available_size[0])
 
         self._win.refresh()
 
     def _layout(self, redraw):
-        '''Propagate layout changes: evaluates the available space and calls
-        layout of the view model.
+        '''Propagate layout changes: evaluates the available space and
+        recompute displayable content from the selected lines.
         '''
+
+        self._ruler = '-' * self._size[1]
 
         # Compute space available for file content
         h, w = self._size
-        self._ruler = '-' * w
         prefix_len, _, _ = self._get_prefix_info()
         h = max(0, h - 1 - self._config.get_filters_count())
         w = max(0, w - prefix_len)
+        self._content_available_size = (h, w)
 
-        self._vm.layout(h, w, self._model.data, self._config.wrap)
-        self._max_visible_lines_count = h
+        self._display = self._selected.layout(h, w, self._config.wrap)
+        self._offsets.layout(h, self._display.lines_count())
 
         # Try preserve current voffset, but make sure we comply with
         # constraints which might just have changed.
-        self.set_v_offset(self._vm.voffset, False)
+        self.set_v_offset(self._offsets.voffset, False)
 
-        assert self._vm.voffset == 0 or \
-            self._vm.voffset < self._vm.lines_count()
+        assert self._offsets.voffset == 0 or \
+            self._offsets.voffset < self._display.lines_count()
 
         if redraw:
             self.draw()
 
     def _sync(self, redraw: bool) -> None:
-        self._model.sync(self._config.filters, self._config.line_visibility)
+        self._selected = self._raw.filter(
+            self._config.filters,
+            self._config.line_visibility)
         self._layout(redraw)
 
     def set_config(self, config: models.ViewConfig) -> None:
@@ -371,7 +304,7 @@ class TextView:
         '''Sets the content of this view. Assumes the view is offscreen and
         does not trigger a redraw.
         '''
-        self._model.set_lines(lines)
+        self._raw.set_lines(lines)
         self._sync(False)
 
     def _pop_filter(self) -> types.Status:
@@ -490,22 +423,22 @@ class TextView:
         return f'Using color palette #{pid}'
 
     def _set_h_offset(self, offset: int) -> types.Status:
-        if self._vm.set_h_offset(offset):
+        if self._offsets.set_h_offset(offset):
             self.draw()
         return STATUS_EMPTY
 
     def set_v_offset(self, offset: int, redraw: int) -> types.Status:
         '''Sets the vertical offset of the view.'''
-        if self._vm.set_v_offset(offset) and redraw:
+        if self._offsets.set_v_offset(offset) and redraw:
             self.draw()
         return STATUS_EMPTY
 
     def _hscroll(self, delta: int) -> types.Status:
-        self._set_h_offset(self._vm.hoffset + delta)
+        self._set_h_offset(self._offsets.hoffset + delta)
         return STATUS_EMPTY
 
     def _vscroll(self, delta: int) -> types.Status:
-        self.set_v_offset(self._vm.voffset + delta, True)
+        self.set_v_offset(self._offsets.voffset + delta, True)
         return STATUS_EMPTY
 
     def goto_line(self, line_text: str) -> types.Status:
@@ -518,12 +451,12 @@ class TextView:
             return 'Not a number'
         if line <= 0:
             self.set_v_offset(0, True)
-        elif line >= self._vm.lines_count():
+        elif line >= self._display.lines_count():
             self.set_v_offset(sys.maxsize, True)
         else:
             voffset = 0
-            for idata, _ in self._vm.data:
-                i, _, _, _ = self._model.data[idata]
+            for iline, _ in self._display.dlines:
+                i, _, _, _ = self._selected.lines[iline]
                 if i >= line:
                     self.set_v_offset(voffset, True)
                     break
@@ -555,22 +488,22 @@ class TextView:
                           starting: bool,
                           direction: int
                           ) -> types.Status:
-        iddata = self._vm.voffset
-        idatamax = len(self._model.data)
-        idata, _ = self._vm.data[iddata]
+        idline = self._offsets.voffset
+        ilinemax = len(self._selected.lines)
+        iline, _ = self._display.dlines[idline]
         if not starting:
-            idata += direction
-        while 0 <= idata < idatamax:
-            _, fidx, _, _ = self._model.data[idata]
+            iline += direction
+        while 0 <= iline < ilinemax:
+            _, fidx, _, _ = self._selected.lines[iline]
             if fidx >= 0:
-                iddata = self._vm.firstdlines[idata]
+                idline = self._display.firstdlines[iline]
                 break
-            idata += direction
+            iline += direction
         # Remember current offset to be able to return a side-effect
         # description
-        prev_offset = self._vm.voffset
-        self.set_v_offset(iddata, True)
-        if prev_offset == self._vm.voffset:
+        prev_offset = self._offsets.voffset
+        self.set_v_offset(idline, True)
+        if prev_offset == self._offsets.voffset:
             return '(END)' if direction > 0 else '(BEGIN)'
         return 'Next match' if direction > 0 else 'Previous match'
 
@@ -586,7 +519,7 @@ class TextView:
         return STATUS_EMPTY
 
     def _vpagescroll(self, delta: int) -> types.Status:
-        self._vscroll(delta * self._max_visible_lines_count)
+        self._vscroll(delta * self._content_available_size[0])
         return STATUS_EMPTY
 
     def execute(self, command: enums.TextViewCommand) -> types.Status:
@@ -699,271 +632,3 @@ class DebugView:
         if len(self._lines) > h:
             self._lines.pop(0)
         self.draw()
-
-
-class Views:
-    '''Aggregates all views, controlling which ones are visible, handling
-    key presses and command dispatching.'''
-
-    # pylint: disable=too-many-instance-attributes
-
-    def __init__(self) -> None:
-        self._debug_view: DebugView
-        self._scr: Any = None
-        self._path: str = ''
-        self._content: List[TextView] = []
-        self._current: int = -1
-        self._hidden_view: int = -1
-        self._y_get_text: int = 0
-        self._mtime: float = 0.0
-        self._auto_reload: bool = False
-        self._auto_reload_anchor: int = 0
-
-    def layout(self) -> types.Size:
-        '''Recompute layout'''
-        assert self._scr
-        scr = self._scr
-        maxh, maxw = get_max_yx(scr)
-        view_lines_count = maxh - 1
-        x = 0
-        y = 0
-
-        if USE_DEBUG:
-            dbg_size = (10, maxw)
-            self._debug_view.layout(dbg_size, (y, x))
-            # Just add padding to expose layout issues in the app
-            padding = 3
-            maxh = max(0, maxh - (2 * padding + dbg_size[0]))
-            maxw = max(0, maxw - 2 * padding)
-            view_lines_count = maxh
-            x = padding
-            y = dbg_size[0] + padding
-
-        for v in self._content:
-            v.layout(view_lines_count, maxw, y, x)
-
-        y += view_lines_count
-        self._y_get_text = y
-        return maxh, maxw
-
-    def _set_view(self, idx: int, propagate_config: bool) -> types.Status:
-        assert 0 <= idx < len(self._content)
-        if self._current != idx:
-            config = self._content[self._current].get_config()
-            self._current = idx
-            if propagate_config:
-                self._content[idx].set_config(copy.deepcopy(config))
-            self._content[idx].show()
-        return f'Switched to {self._content[idx].name()}'
-
-    def _help_view_push(self) -> types.Status:
-        self._hidden_view = self._current
-        return self._set_view(3, False)
-
-    def _help_view_pop(self) -> types.Status:
-        idx = self._hidden_view
-        self._hidden_view = -1
-        return self._set_view(idx, False)
-
-    def _reload(self, scroll_to: int) -> types.Status:
-        with open(self._path, encoding='utf-8') as f:
-            lines = f.readlines()
-            for i, v in enumerate(self._content):
-                if i == len(self._content) - 1:
-                    v.set_lines(HELP_LINES)
-                else:
-                    v.set_lines(lines)
-                    v.set_v_offset(scroll_to, False)
-        self._content[self._current].draw()
-        self._mtime = getmtime(self._path)
-        return 'File reloaded'
-
-    def create(self, store, scr, path: str) -> None:
-        '''Creates all views in the given screen, and loads the content from
-        the given file.'''
-        self._scr = scr
-        self._path = path
-        self._content.clear()
-        self._content.append(TextView(store, scr, 'View 1', path))
-        self._content.append(TextView(store, scr, 'View 2', path))
-        self._content.append(TextView(store, scr, 'View 3', path))
-        self._content.append(TextView(store, scr, 'Help', 'Help'))
-        if USE_DEBUG:
-            self._debug_view = DebugView(scr)
-            debug.OUT_FUNC = self._debug_view.out
-        self.layout()
-        self._reload(0)
-        self._set_view(0, False)
-
-    def prompt(self, text_prompt: str, text: str) -> str:
-        '''Prompts user to enter or edit some text.'''
-        return prompt(self._scr, self._y_get_text, 0, text_prompt, text)
-
-    def _get_keyword(self) -> str:
-        return self.prompt('Keyword: ', '')
-
-    def try_start_search(self) -> types.Status:
-        '''Tries to initiate a less like search.'''
-        v = self._content[self._current]
-        if v.has_filters():
-            return 'Cannot start search when filters already defined'
-        t = self._get_keyword()
-        return v.search(t)
-
-    def _toggle_auto_reload(self, anchor: int):
-        self._auto_reload = not self._auto_reload
-        self._auto_reload_anchor = anchor
-        return f'Auto reload {self._auto_reload}'
-
-    def _poll(self) -> Tuple[bool, types.Status]:
-        if self._auto_reload:
-            mtime = getmtime(self._path)
-            if mtime != self._mtime:
-                return True, self._reload(self._auto_reload_anchor)
-        return False, STATUS_UNCHANGED
-
-    def handle_key(self, key) -> Tuple[bool, types.Status]:
-        '''Handles the given key, propagating it to the proper view.'''
-
-        if key == -1:
-            # No key, means we are polling view for any self triggered action
-            return self._poll()
-
-        # Local functions redirecting to current view v
-        v = self._content[self._current]
-
-        if key == curses.KEY_RESIZE:
-            self._scr.clear()
-            self._scr.refresh()
-            size = self.layout()
-            v.draw()
-            return True, f'Resized to {size[1]}x{size[0]}'
-
-        def new_keyword(new_filter) -> types.Status:
-            keyword = self._get_keyword()
-            return v.push_keyword(keyword, new_filter)
-
-        def goto_line() -> types.Status:
-            line_as_text = self.prompt('Enter line: ', '')
-            return v.goto_line(line_as_text)
-
-        def edit_keyword() -> types.Status:
-            count, keyword = v.get_last_keyword()
-            if count <= 0:
-                return 'No keyword to edit'
-            keyword = self.prompt('Edit: ', keyword if keyword else '')
-            if len(keyword) <= 0:
-                return 'No change made'
-            v.execute(enums.TextViewCommand.POP_KEYWORD)
-            v.push_keyword(keyword, count == 1)
-            return 'Keyword updated'
-
-        # Map keys to simple text view commands that don't take any
-        # argument and that can be directly handled by calling v.execute()
-        keys_to_command = {
-            ord('F'):              enums.TextViewCommand.POP_FILTER,
-            curses.KEY_BACKSPACE:  enums.TextViewCommand.POP_FILTER,
-            ord('-'):              enums.TextViewCommand.POP_KEYWORD,
-            ord('_'):              enums.TextViewCommand.POP_KEYWORD,
-            ord('n'):              enums.TextViewCommand.VSCROLL_TO_NEXT_MATCH,
-            ord('p'):              enums.TextViewCommand.VSCROLL_TO_PREV_MATCH,
-            ord('l'):              enums.TextViewCommand.TOGGLE_LINE_NUMBERS,
-            ord('k'):              enums.TextViewCommand.TOGGLE_WRAP,
-            ord('*'):              enums.TextViewCommand.TOGGLE_BULLETS,
-            ord('.'):              enums.TextViewCommand.TOGGLE_SHOW_SPACES,
-            ord('c'):              enums.TextViewCommand.NEXT_PALETTE,
-            ord('C'):              enums.TextViewCommand.PREV_PALETTE,
-            ord('h'):              enums.TextViewCommand.NEXT_COLORIZE_MODE,
-            ord('H'):              enums.TextViewCommand.PREV_COLORIZE_MODE,
-            ord('m'):              enums.TextViewCommand.NEXT_LINE_VISIBILITY,
-            ord('M'):              enums.TextViewCommand.PREV_LINE_VISIBILITY,
-            ord('i'):              enums.TextViewCommand.TOGGLE_IGNORE_CASE,
-            ord('x'):              enums.TextViewCommand.TOGGLE_HIDING,
-            curses.KEY_UP:         enums.TextViewCommand.GO_UP,
-            curses.KEY_DOWN:       enums.TextViewCommand.GO_DOWN,
-            curses.KEY_LEFT:       enums.TextViewCommand.GO_LEFT,
-            curses.KEY_RIGHT:      enums.TextViewCommand.GO_RIGHT,
-            curses.KEY_HOME:       enums.TextViewCommand.GO_HOME,
-            ord('g'):              enums.TextViewCommand.GO_HOME,
-            ord('<'):              enums.TextViewCommand.GO_HOME,
-            curses.KEY_END:        enums.TextViewCommand.GO_END,
-            ord('G'):              enums.TextViewCommand.GO_END,
-            ord('>'):              enums.TextViewCommand.GO_END,
-            curses.KEY_NPAGE:      enums.TextViewCommand.GO_NPAGE,
-            ord(' '):              enums.TextViewCommand.GO_NPAGE,
-            curses.KEY_PPAGE:      enums.TextViewCommand.GO_PPAGE,
-            ord('b'):              enums.TextViewCommand.GO_PPAGE,
-            curses.KEY_SLEFT:      enums.TextViewCommand.GO_SLEFT,
-            curses.KEY_SRIGHT:     enums.TextViewCommand.GO_SRIGHT,
-            ord('d'):              enums.TextViewCommand.SWAP_FILTERS,
-            ord('w'):              enums.TextViewCommand.ROTATE_FILTERS_UP,
-            ord('s'):              enums.TextViewCommand.ROTATE_FILTERS_DOWN,
-            ord('\\'):             enums.TextViewCommand.SLOT_SAVE,
-            ord('|'):              enums.TextViewCommand.SLOT_DELETE,
-            ord(']'):              enums.TextViewCommand.SLOT_LOAD_NEXT,
-            ord('['):              enums.TextViewCommand.SLOT_LOAD_PREV,
-        }
-
-        # Map keys to custom functions used to handle more complex commands
-        # like ones taking arguments
-
-        # noqa: E272 is "multiple spaces before keyword"
-        keys_to_func = {
-            ord('1'):
-                lambda: self._set_view(0, False),
-            ord('2'):
-                lambda: self._set_view(1, False),
-            ord('3'):
-                lambda: self._set_view(2, False),
-            ord('!'):
-                lambda: self._set_view(0, True),
-            ord('@'):
-                lambda: self._set_view(1, True),
-            ord('#'):
-                lambda: self._set_view(2, True),
-            ord('?'):
-                self._help_view_push,
-            ord('e'):
-                edit_keyword,
-            ord('f'):
-                lambda: new_keyword(True),
-            ord('\n'):
-                lambda: new_keyword(True),
-            ord('r'):
-                lambda: self._reload(0),
-            ord('R'):
-                lambda: self._toggle_auto_reload(0),
-            ord('t'):
-                lambda: self._reload(sys.maxsize),
-            ord('T'):
-                lambda: self._toggle_auto_reload(sys.maxsize),
-            ord('+'):
-                lambda: new_keyword(False),
-            ord('='):
-                lambda: new_keyword(False),
-            ord('/'):
-                self.try_start_search,
-            curses.ascii.TAB:
-                goto_line,
-            7:
-                goto_line,
-        }
-
-        inter = keys_to_command.keys() & keys_to_func.keys()
-        assert len(inter) == 0, f'Keys with multiple mapping {inter}'
-        status: types.Status = ''
-
-        # If help is shown, we hijack keys closing the view but
-        # forward all other keys as if it is a regular view (which
-        # makes help searchable like a file...)
-        if self._hidden_view >= 0 and key in (
-                ord('q'), ord('Q'), curses.ascii.ESC):
-            self._help_view_pop()
-        elif key in keys_to_command:
-            status = v.execute(keys_to_command[key])
-        elif key in keys_to_func:
-            status = types.Status(keys_to_func[key]())
-        else:
-            return False, f'Unknown key {key} (? for help, q to quit)'
-
-        return True, status
