+++ setup.cfg
@@ -17,5 +17,5 @@ install_requires =
 * = *.md, *.txt
 [options.entry_points]
 console_scripts:
-    searchf=searchf.app:main
+    searchf=searchf.main:main
     searchf-test=searchf.test.all:main
